doc/internals/: Fix issues with CHARACTER description.
  * A couple of bits here.  First, I managed to describe the size
of the various fields in the character object correctly, but then
described their position as though the top part of the character
code was at the top of a 64-bit word, effectively extending the
range for the code points by about 32 bits.  Oops.  Second, I got
confirmation that the code points are definitely all unicode.
Fixed.